Fixing specs
